+++ i-make/libs/jsonc.py
@@ -1,0 +1,10 @@
+import json
+import re
+
+
+def load_jsonc(file_path: str):
+    with open(file_path, "r", encoding="utf-8") as f:
+        text = f.read()
+    re_text = re.sub(r"/\*[\s\S]*?\*/|//.*", "", text)
+    json_obj = json.loads(re_text)
+    return json_obj
